extended hedge mode
